https: client request doesn’t define default encoding (#440)
This shares the client class at http/https implementations.
