+++ src/js/https_client.js
@@ -16,255 +16,13 @@
 
 var util = require('util');
 var tls = require('tls');
-var stream = require('stream');
-var OutgoingMessage = require('http_outgoing').OutgoingMessage;
-var common = require('http_common');
-var HTTPParser = require('httpparser').HTTPParser;
+var HTTPClientRequest = require('http_client').ClientRequest;
 
 function ClientRequest(options, cb) {
-  OutgoingMessage.call(this);
-
-  this.aborted = undefined;
-
-  // get port, host and method.
-  var port = options.port = options.port || 443;
-  var host = options.host = options.hostname || options.host || '127.0.0.1';
-  var method = options.method || 'GET';
-  var path = options.path || '/';
-
-  // buffer for cached..
-  this._buffer = null;
-
-  // If `options` contains header information, save it.
-  if (options.headers) {
-    var keys = Object.keys(options.headers);
-    for (var i = 0, l = keys.length; i < l; i++) {
-      var key = keys[i];
-      this.setHeader(key, options.headers[key]);
-    }
-  }
-
-  if (host && !this.getHeader('host')) {
-    var hostHeader = host;
-    if (port && +port !== 443) {
-      hostHeader += ':' + port;
-    }
-    this.setHeader('Host', hostHeader);
-  }
-
-  // store first header line to be sent.
-  this._storeHeader(method + ' ' + path + ' HTTP/1.1\r\n');
-
-  // Register response event handler.
-  if (cb) {
-    this.once('response', cb);
-  }
-
-
-  // Create socket.
-  var socket = tls.connect(Object.assign(options, {
-    host: host,
-    port: port,
-  }));
-
-  // setup connection information.
-  setupConnection(this, socket);
+  options.Socket = tls.TLSSocket;
+  options.port = options.port || 443;
+  HTTPClientRequest.call(this, options, cb);
 }
-
-util.inherits(ClientRequest, OutgoingMessage);
+util.inherits(ClientRequest, HTTPClientRequest);
 
 exports.ClientRequest = ClientRequest;
-
-
-function setupConnection(req, socket) {
-  var parser = common.createHTTPParser();
-  parser.reinitialize(HTTPParser.RESPONSE);
-  socket.parser = parser;
-  socket._httpMessage = req;
-
-  parser.socket = socket;
-  parser.incoming = null;
-  parser._headers = [];
-  parser.onIncoming = parserOnIncomingClient;
-
-  req.socket = socket;
-  req.connection = socket;
-  req.parser = parser;
-
-  socket.on('error', socketOnError);
-  socket.on('data', socketOnData);
-  socket.on('end', socketOnEnd);
-  socket.on('close', socketOnClose);
-
-  // socket emitted when a socket is assigned to req
-  process.nextTick(function() {
-    req.emit('socket', socket);
-  });
-}
-
-function cleanUpSocket(socket) {
-  var parser = socket.parser;
-  var req = socket._httpMessage;
-
-  if (parser) {
-    // unref all links to parser, make parser GCed
-    parser.finish();
-    parser = null;
-    socket.parser = null;
-    req.parser = null;
-  }
-
-  socket.destroy();
-}
-
-function emitError(socket, err) {
-  var req = socket._httpMessage;
-
-  if (err) {
-    var host;
-    if (host = req.getHeader('host')) {
-      err.message += ': ' + (host ? host : '');
-    }
-    req.emit('error', err);
-  }
-}
-
-function socketOnClose() {
-  var socket = this;
-  var req = socket._httpMessage;
-
-  // socket.read();
-
-  req.emit('close');
-
-  if (req.res && req.res.readable) {
-    // Socket closed before we emitted 'end' below.
-    if (!req.res.complete) req.res.emit('aborted');
-    // Socket closed before we emitted 'end'
-    var res = req.res;
-    res.on('end', function() {
-      res.emit('close');
-    });
-    res.push(null);
-  }
-
-  cleanUpSocket(this);
-}
-
-function socketOnError(err) {
-  cleanUpSocket(this);
-  emitError(this, err);
-}
-
-function socketOnData(d) {
-  var socket = this;
-  var req = this._httpMessage;
-  var parser = this.parser;
-
-  if (!this._buffer) {
-    this._buffer = d;
-  } else {
-    this._buffer = Buffer.concat([this._buffer, d]);
-  }
-
-  if (this._buffer.valid('utf8')) {
-    ondata(this._buffer);
-    this._buffer = null;
-  }
-
-  function ondata(valid) {
-    var ret = parser.execute(valid);
-    if (ret instanceof Error) {
-      cleanUpSocket(socket);
-      req.emit('error', ret);
-    }
-  }
-}
-
-function socketOnEnd() {
-  cleanUpSocket(this);
-}
-
-// This is called by parserOnHeadersComplete after response header is parsed.
-// TODO: keepalive support
-function parserOnIncomingClient(res, shouldKeepAlive) {
-  var socket = this.socket;
-  var req = socket._httpMessage;
-
-  if (req.res) {
-    // server sent responses twice.
-    socket.destroy();
-    return false;
-  }
-  req.res = res;
-
-  res.req = req;
-
-  res.on('end', responseOnEnd);
-
-  req.emit('response', res);
-
-  // response to HEAD req has no body
-  var isHeadResponse = (req.method === 'HEAD');
-
-  return isHeadResponse;
-
-}
-
-function responseOnEnd() {
-  var res = this;
-  var req = res.req;
-  var socket = req.socket;
-
-  if (socket._socketState.writable) {
-    socket.destroySoon();
-  }
-}
-
-ClientRequest.prototype.abort = function() {
-  var self = this;
-  if (!self.aborted) {
-    process.nextTick(function() {
-      self.emit('abort');
-    });
-  }
-  // Mark as aborting so we can avoid sending queued request data
-  // This is used as a truthy flag elsewhere. The use of Date.now is for
-  // debugging purposes only.
-  self.aborted = Date.now();
-
-  // If we're aborting, we don't care about any more response data.
-  if (self.res) {
-    self.res._dump();
-  } else {
-    self.once('response', function(res) {
-      res._dump();
-    });
-  }
-
-  // the request queue through handling in onSocket.
-  if (self.socket) {
-    // in-progress
-    self.socket._writableState.buffer = [];
-    stream.Duplex.prototype.end.call(self.socket, undefined, function() {
-      self.socket.destroy();
-    });
-  }
-};
-
-ClientRequest.prototype.setTimeout = function(ms, cb) {
-  var self = this;
-
-  if (cb) self.once('timeout', cb);
-
-  function emitTimeout() {
-    self.emit('timeout');
-  }
-
-  // In IoT.js, socket is already assigned,
-  // thus, it is sufficient to trigger timeout on socket 'connect' event.
-  this.socket.once('connect', function() {
-    self.socket.setTimeout(ms, emitTimeout);
-  });
-
-};
